test: split app interactions coverage
